+++ app/views/tables/cells/linked-card.js
@@ -3,7 +3,7 @@ import Ember from "ember";
 var LinkedCardView = Ember.View.extend({
 	tagName: "div",
 	templateName: "tables/cells/linked-card",
-	classNameBindings: [":card", ":col-md-3", "item.isOverdue:overdue"],
+	classNameBindings: [":card", ":col-md-3"],
 });
 
 export default LinkedCardView;